Trying to optimize: int -> byte

+++ src/Config.java
@@ -1,5 +1,5 @@
 class Config {
     static int size = 4;
-    static int[][] solvedTiles = TileGenerator.generate(BoardState.SOLVED, size);
+    static byte[][] solvedTiles = TileGenerator.generate(BoardState.SOLVED, size);
     static IMetrics metrics = new Manhattan(solvedTiles);
 }
